+++ src/hocs/interfaces/IWithList.ts
@@ -1,0 +1,6 @@
+export interface IListProps<T> {
+  Component: React.ComponentType<T>;
+  id?: string;
+  data: T[];
+  res: object;
+}