+++ src/modules/episodes/interfaces/IEpisodeTable.ts
@@ -1,0 +1,12 @@
+export interface IData {
+  episode: string;
+  name: string;
+  air_date: string;
+}
+
+export interface IEpisodeTable {
+  data: IData[];
+  isLoading?: boolean;
+  title: string;
+  id: string;
+}